nytimes_microservice/requirements.txt
# FastAPI Web Framework
fastapi==0.110.2

# ASGI Server
uvicorn[standard]==0.29.0

# HTTP Client
httpx==0.27.0

# Data Validation
pydantic==2.7.1

# Environment Variable Loader
python-dotenv==1.0.1

# Testing
pytest==8.2.2
pytest-asyncio==0.23.6

# Optional (Dev Tools)
ruff==0.4.5
mypy==1.10.0